Output and Tkinter Grid
Placed a textbox widget in Tkinter which shows the output of the
Hothand. Also restructured Tkinter into the grid format.
